系统初始化模块完成

+++ src/main/java/com/kevin/mem/mng/dto/response/KeyValueLevelDTO.java
@@ -11,7 +11,7 @@ import java.util.List;
  */
 @Data
 public class KeyValueLevelDTO {
-    private Integer code;//字段编号
+    private Long code;//字段编号
 
     private String desc;//字段描述
